Networking Changes [Not finished]
Added functionality for the client to be able to change the conversation list of the chatbot. This can allow for saving and resuming a conversation. server.py has a function that performs the change but will not work as it is taking a string not an array object. I needs to process the string.