Fixed issue with rendering
Finally fully fixed the issue with the sprite not rendering correctly. This issue was caused by the windows scale setting. After some exponential statistical analysis and using a formula to compensate for scaling, there are no more boxes. For the sprite to properly sit on the taskbar, a linear formula was created that also compensates for scaling to maintain the established offset.