Animated button
Animated Button made (the one that works is in file called backup.py), made it so if you change one of the parameters the animation becomes the hover one. Now I have to use that logic to make the entire UI cuz the tutorial did some weird thing with the beginning self and root stuff idk.