Fixed loading prev convo and program start delay
Fixed how the server sent the conversation. Since the training text has Action: and Code:, this inadvertently messed up how processing the response worked when sending the conversation. This was fixed on the server side code. Also added multithreading for the client being initialized in main as it would result in long boot time for the sprite. The client now loads in the background on a separate thread which begins before the preprocessing of animations (another significant time use). This concurrency maximizes the efficiency of the program.